feat(#184): split refresh in client
